Cast magma_int_t values to (int) in printf, to avoid warnings when 64-bit. Delete some unused variables. (in interface)
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@3082 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ interface_cuda/error.cpp
@@ -40,7 +40,7 @@ void magma_xerror( magma_err_t err, const char* func, const char* file, int line
 {
     if ( err != MAGMA_SUCCESS ) {
         fprintf( stderr, "MAGMA error: %s (%d) in %s at %s:%d\n",
-                 magma_strerror( err ), err, func, file, line );
+                 magma_strerror( err ), (int) err, func, file, line );
     }
 }
